Add additional metadata files which are needed for our docker image file builder setup.
Update affected tests.
