+++ tests/distribution/python_version_change_tests/amazonlinux2_test.py
@@ -18,7 +18,7 @@ from __future__ import print_function
 from __future__ import absolute_import
 import pytest
 
-from tests.distribution_builders.centos import CentOSBuilder
+from tests.distribution_builders.amazonlinux2 import AmazonlinuxBuilder
 
 from tests.utils.dockerized import dockerized_case
 from tests.distribution.python_version_change_tests.common import (
@@ -39,8 +39,8 @@ from tests.utils.agent_runner import AgentRunner, PACKAGE_INSTALL
 
 
 @pytest.mark.usefixtures("agent_environment")
-@dockerized_case(CentOSBuilder, __file__)
-def test_centos_test_versions(request):
+@dockerized_case(AmazonlinuxBuilder, __file__)
+def test_amazonlinux_test_versions(request):
     runner = AgentRunner(PACKAGE_INSTALL)
     common_version_test(
         runner,
@@ -53,14 +53,7 @@ def test_centos_test_versions(request):
         install_fails=True,
     )
     common_version_test(
-        runner,
-        install_rpm,
-        remove_rpm,
-        "config_main.py",
-        "",
-        "2.5.1",
-        "3.4.1",
-        install_fails=True,
+        runner, install_rpm, remove_rpm, "config_main.py", "", "2.5.1", "3.4.1"
     )
     common_version_test(
         runner, install_rpm, remove_rpm, "config_main_py2.py", "2.5.1", "", "3.4.1"
@@ -68,67 +61,65 @@ def test_centos_test_versions(request):
     common_version_test(
         runner, install_rpm, remove_rpm, "config_main_py3.py", "2.5.1", "2.5.1", ""
     )
-    common_version_test(
-        runner, install_rpm, remove_rpm, "config_main_py2.py", "", "", ""
-    )
+    common_version_test(runner, install_rpm, remove_rpm, "config_main.py", "", "", "")
     common_version_test(
         runner, install_rpm, remove_rpm, "config_main_py2.py", "2.5.1", "", ""
     )
     common_version_test(
-        runner, install_rpm, remove_rpm, "config_main_py3.py", "", "2.5.1", ""
+        runner, install_rpm, remove_rpm, "config_main.py", "", "2.5.1", ""
     )
 
 
 @pytest.mark.usefixtures("agent_environment")
-@dockerized_case(CentOSBuilder, __file__)
-def test_centos_switch_command_works_without_agent_config(request):
+@dockerized_case(AmazonlinuxBuilder, __file__)
+def test_amazonlinux_switch_command_works_without_agent_config(request):
     common_test_switch_command_works_without_agent_config(install_rpm)
 
 
 @pytest.mark.usefixtures("agent_environment")
-@dockerized_case(CentOSBuilder, __file__)
-def test_centos_no_python(request):
+@dockerized_case(AmazonlinuxBuilder, __file__)
+def test_amazonlinux_no_python(request):
     common_test_no_python(install_rpm)
 
 
 @pytest.mark.usefixtures("agent_environment")
-@dockerized_case(CentOSBuilder, __file__)
-def test_centos_only_python_mapped_to_python2(request):
+@dockerized_case(AmazonlinuxBuilder, __file__)
+def test_amazonlinux_only_python_mapped_to_python2(request):
     common_test_only_python_mapped_to_python2(install_rpm, install_next_version_rpm)
 
 
 @pytest.mark.usefixtures("agent_environment")
-@dockerized_case(CentOSBuilder, __file__)
-def test_centos_only_python_mapped_to_python3(request):
+@dockerized_case(AmazonlinuxBuilder, __file__)
+def test_amazonlinux_only_python_mapped_to_python3(request):
     common_test_only_python_mapped_to_python3(install_rpm, install_next_version_rpm)
 
 
 @pytest.mark.usefixtures("agent_environment")
 @pytest.mark.timeout(30)
-@dockerized_case(CentOSBuilder, __file__)
-def test_centos_python2(request):
+@dockerized_case(AmazonlinuxBuilder, __file__)
+def test_amazonlinux_python2(request):
     common_test_python2(install_rpm, install_next_version_rpm)
 
 
 @pytest.mark.usefixtures("agent_environment")
-@dockerized_case(CentOSBuilder, __file__)
-def test_centos_python3(request):
+@dockerized_case(AmazonlinuxBuilder, __file__)
+def test_amazonlinux_python3(request):
     common_test_python3(install_rpm, install_next_version_rpm)
 
 
 @pytest.mark.usefixtures("agent_environment")
-@dockerized_case(CentOSBuilder, __file__)
-def test_centos_switch_default_to_python2(request):
+@dockerized_case(AmazonlinuxBuilder, __file__)
+def test_amazonlinux_switch_default_to_python2(request):
     common_test_switch_default_to_python2(install_rpm, install_next_version_rpm)
 
 
 @pytest.mark.usefixtures("agent_environment")
-@dockerized_case(CentOSBuilder, __file__)
-def test_centos_switch_default_to_python3(request):
+@dockerized_case(AmazonlinuxBuilder, __file__)
+def test_amazonlinux_switch_default_to_python3(request):
     common_test_switch_default_to_python3(install_rpm, install_next_version_rpm)
 
 
 @pytest.mark.usefixtures("agent_environment")
-@dockerized_case(CentOSBuilder, __file__)
-def test_centos_switch_python2_to_python3(request):
+@dockerized_case(AmazonlinuxBuilder, __file__)
+def test_amazonlinux_switch_python2_to_python3(request):
     common_test_switch_python2_to_python3(install_rpm, install_next_version_rpm)
