+++ tests/distribution/python_version_change_tests/ubuntu1804_test.py
@@ -18,7 +18,7 @@ from __future__ import print_function
 from __future__ import absolute_import
 import pytest
 
-from tests.distribution_builders.ubuntu import UbuntuBuilder
+from tests.distribution_builders.ubuntu1804_with_py3 import UbuntuBuilder
 
 from tests.utils.dockerized import dockerized_case
 from tests.distribution.python_version_change_tests.common import (
