+++ tests/distribution_builders/centos8/cmd.py
@@ -16,7 +16,7 @@ from __future__ import print_function
 from __future__ import absolute_import
 from __future__ import unicode_literals
 
-from tests.distribution_builders.centos import CentOSBuilder
+from tests.distribution_builders.centos8 import CentOSBuilder
 
 if __name__ == "__main__":
     CentOSBuilder.handle_command_line()
